requirements.txt
\
chardet>=5.1
llama-cpp-python>=0.2.0; platform_system != "Emscripten" and platform_system != "iOS"
